fixed the navbar
