+++ apps/web/app/(authenticated)/mentorship/page.tsx
@@ -1,4 +1,4 @@
-import NavBar from "../lib/components/NavBar/navbar";
+import NavBar from "../lib/components/NavBar/nav-bar";
 export default function Mentorship() {
   return (
     <div>
