+++ apps/web/app/(unauthenticated)/layout.tsx
@@ -1,0 +1,14 @@
+import WelcomeBar from "../lib/components/NavBar/welcome-bar";
+
+export default function UnauthenticatedLayout({
+    children,
+}: Readonly<{
+    children: React.ReactNode;
+}>) {
+    return (
+        <>
+            <WelcomeBar />
+            {children}
+        </>
+    );
+}
