+++ apps/web/app/(unauthenticated)/login/page.tsx
@@ -1,18 +1,12 @@
 'use client'
 
 import Link from "next/link"
-
 import { Input } from "@repo/ui/components/ui/input"
 import { Button } from "@repo/ui/components/ui/button"
 import { Mail, Lock } from "lucide-react"
 
-
 import { useRouter } from "next/navigation"
-import { useFormState } from "react-dom"
-import { verifyEmail, verifyEmailState } from "./VerifyEmail"
 import { useState } from "react"
-
-
 
 export default function LoginPage() {
   const router = useRouter();
