+++ apps/web/app/(unauthenticated)/page.tsx
@@ -1,4 +1,3 @@
-"use client"; 
 import React from "react";
 
 const WelcomeSection: React.FC = () => {
