+++ apps/web/app/lib/components/NavBar/notifications-list.tsx
@@ -92,12 +92,11 @@ export async function NotificationsList() {
             <div key={notification.notifID} className="flex items-center justify-between gap-4 py-2">
               <div className="flex flex-col gap-1">
                 {/* <span className="font-medium">{notification.username}</span> */}
-                <button
-                  onClick={() => console.log(`Navigate to profile ${notification.notifID}`)}
+                <Button
                   className="font-medium text-left hover:underline cursor-pointer"
                 >
                   {notification.userEmail}
-                </button>
+                </Button>
 
                 <span className="text-sm text-muted-foreground">{notification.message}</span>
               </div>
