+++ apps/web/app/lib/components/Onboarding/BioStep.tsx
@@ -3,7 +3,7 @@ import { Label } from "@repo/ui/components/ui/label";
 import { Button } from "@repo/ui/components/ui/button";
 import { Textarea } from "@repo/ui/components/ui/textarea";
 import { FileText, ArrowLeft, ArrowRight } from "lucide-react";
-import type { OnboardingData } from "@/app/signup/onboarding";
+import type { OnboardingData } from "@/app/(unauthenticated)/signup/onboarding";
 
 interface BioStepProps {
   onNext: (data: { bio: string }) => void;
