+++ apps/web/app/lib/components/Onboarding/CredentialsStep.tsx
@@ -3,7 +3,7 @@ import { Label } from "@repo/ui/components/ui/label";
 import { Button } from "@repo/ui/components/ui/button";
 import { Input } from "@repo/ui/components/ui/input";
 import { Lock, UserPenIcon, ArrowLeft, ArrowRight } from "lucide-react";
-import type { OnboardingData } from "@/app/signup/onboarding";
+import type { OnboardingData } from "@/app/(unauthenticated)/signup/onboarding";
 
 interface CredentialsStepProps {
   onNext: (data: Partial<OnboardingData>) => void;
