+++ apps/web/app/lib/components/Onboarding/ResumeUploadStep.tsx
@@ -5,7 +5,7 @@ import { Button } from "@repo/ui/components/ui/button";
 import { Input } from "@repo/ui/components/ui/input";
 import { Label } from "@repo/ui/components/ui/label";
 import { ArrowLeft, ArrowRight } from "lucide-react";
-import type { OnboardingData } from "../../../signup/onboarding";
+import type { OnboardingData } from "../../../(unauthenticated)/signup/onboarding";
 // import { uploadResume } from "@/app/lib/api";
 
 interface ResumeUploadStepProps {
