--- apps/web/app/page.tsx
@@ -1,9 +1,0 @@
-import WelcomeSection from "@/app/lib/components/WelcomePage";
-import WelcomeBar from "./lib/components/welcomebar";
-export default function Page() {
-  return (
-    <div>
-      <WelcomeSection />
-    </div>
-  );
-}
